feat: add AppLayout and TopBar components for game session management

+++ backend/blue-route/src/main/java/at/fhv/blueroute/ship/application/service/ShipService.java
@@ -226,8 +226,8 @@ public class ShipService {
         );
 
         ship.setCargoCapacity(offer.getType().getCapacity());
-        ship.setCondition(offer.getCondition());
-        ship.setFuelLevel(offer.getFuelLevel());
+        ship.setCondition((int) offer.getCondition());
+        ship.setFuelLevel((int) offer.getFuelLevel());
         ship.setCurrentPort(player.getCurrentPort());
 
         playerRepository.save(player);
